builder.prepare: memoize results as long as the `Builder` lives
Dropped the `settings` parameter, as it couldn't be cached (dict isn't
hashable) and wasn't used anyway.  If this become an issue, I could
special-case the hashing of `dict`s in `scoped_cache`

This avoids rebuilds with successive method calls.

The results are not meaningful anymore outside a given build's context,
and `scoped_cache` ties the cached outputs to the `Builder` object so
we cannot leak memory.